Update classify_2.py
Updated Classify_2 with several changes.  There was an error in the ant_to_rec_dict (str changed to int); updated fields (this is just basic);  Added lines 62-69 to establish tblTrain as the training dataset.  Weird that this wasn't there before.  Then updated line 72 to specify the tblTrain dataframe;  Finally altered line 80 so that ant-to-rec-dict gets passed as a list instead of a string