Update __init__.py
Updated to retain antenna number in tables--although this adds a field, it is important when antenna switching occurs, specifically if the antennas are grouped togther for classification, but we wish to retain the information for directionality, etc.  Note changes made to Lotek and Orion import routines.  Have only tested on Lotek SRX800