Update __init__.py
Modified bouts because there was an issue with the lag being classified as int32.  I notice there was subsecond info in there (unclear why) so I changed it to float64.  Also noticed a syntax issue with the directory being called as inputWS when it's really coming from scratchWS so I changed that in init.py and also some tweaks to bouts.